refactor!: unified `SpringSimulationController` interface to match `AnimationController` more closely
Adds `.forward()` and `.reverse()` methods, which animate to `upperBound` and `lowerBound` respectively, as well as a `.resync(TickerProvider)` method.

BREAKING CHANGE: The default constructors in `SpringSimulationController` and `SpringSimulationController2D` are no longer unbounded to match `AnimationController` more closely. Use the `.unbounded` in places where you didn't set a `lowerBound` or `upperBound`.