feat!: add `MotionConverter` API to simplify the usage of controllers
This change adds a `MotionConverter` class that allows you to easily create
`MotionController`s for all kinds of different value types.

The `MotionConverter` class is a simple class that converts any value of any
dimensionality so that it can be used with the `MotionController`s.

This makes it easy for example to create a `MotionController<Offset>`,
`MotionController<Size>` or even `MotionController<Color>` etc.